+++ main.go
@@ -27,7 +27,6 @@ import (
 	"os"
 	"os/signal"
 	"path"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -220,51 +219,8 @@ func populateArgs(c *cli.Context) (
 	return
 }
 
-func getResolvedPath(filePath string) (resolvedPath string, err error) {
-	if path.IsAbs(filePath) {
-		return filePath, nil
-	}
-
-	if strings.HasPrefix(filePath, "~/") {
-		homeDir, err1 := os.UserHomeDir()
-		if err1 != nil {
-			return "", fmt.Errorf("get home dir: %w", err1)
-		}
-		return filepath.Join(homeDir, filePath[2:]), nil
-	} else {
-		return filepath.Abs(filePath)
-	}
-}
-
-func resolveRelativePath(c *cli.Context) (err error) {
-	parentProcessExecutionDir, ok := os.LookupEnv("parent_process_execution_dir")
-	if !ok { // Don't do anything, directory is not set
-		return nil
-	}
-	//var parentProcessExecutionDir string
-	logFile := c.String("log-file")
-
-	if logFile == "" || path.IsAbs(logFile) {
-		return nil
-	} else if strings.HasPrefix(logFile, "~") {
-		var resolvedPath string
-		resolvedPath, err = getResolvedPath(logFile)
-		if err != nil {
-			return fmt.Errorf("while resolving path: %w", err)
-		}
-		c.Set("log-file", resolvedPath)
-	} else { // relative to parent process's execution directory
-		err = c.Set("log-file", filepath.Join(parentProcessExecutionDir, logFile))
-		if err != nil {
-			return fmt.Errorf("while setting value in context: %w", err)
-		}
-	}
-
-	return nil
-}
-
 func runCLIApp(c *cli.Context) (err error) {
-	err = resolveRelativePath(c)
+	err = resolvePathInSubprocessForTheRequiredFlag(c)
 	if err != nil {
 		return fmt.Errorf("Resolving path: %w", err)
 	}
@@ -348,9 +304,9 @@ func runCLIApp(c *cli.Context) (err error) {
 		if err != nil {
 			return fmt.Errorf("Getting current working dir: %w", err)
 		}
-		env = append(env, fmt.Sprintf("parent_process_execution_dir=%s", parentProcessExecutionDir))
+		env = append(env, fmt.Sprintf("%s=%s", PARENT_PROCESS_EXECUTION_DIR_KEY, parentProcessExecutionDir))
 
-		if strings.HasPrefix(flags.LogFile, "~") {
+		if strings.HasPrefix(flags.LogFile, "~") || strings.HasPrefix(flags.KeyFile, "~") {
 			// pass home directory, it is not set for sub-process
 			var homeDir string
 			homeDir, err = os.UserHomeDir()
